menu/pgpatrol/requirements.txt
requests~=2.18.4
python-decouple==3.1
